refactoring code
add controllers
